requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
web3==6.11.3
gql==3.4.1
aiohttp==3.9.1
asyncio==3.4.3
scikit-learn==1.3.2 